fix: resolve 17 post-Spring-removal bugs for custom IoC module loading
After replacing Spring with SimpleContainer IoC, module plugins failed to
load due to cascading issues across the framework. All 11 modules now load
successfully with zero errors.

Framework fixes:
- SimpleContainer.getBean(): add assignability fallback for interface-typed deps
- Register UltiTools instance in core context so Plugin.class is resolvable
- ComponentScanner: support jar: protocol (modules are JARs, not directories)
- AnnotationUtils: add meta-annotation search (@UltiToolsModule -> @EnableAutoRegister)
- SimpleContainer.createBean(): add constructor auto-wiring fallback
- ConfigManager.registerAll(): try no-arg constructor when String(path) fails
- Register config entities as beans for @Autowired injection
- PluginManager: set static instance via reflection before refresh() so
  @PostConstruct methods can call getInstance().getDataOperator()
- PluginManager: use loadClass() instead of loadPluginClass() during JAR
  scanning to avoid false SecurityException warnings on non-plugin classes
- CommandManager/ListenerManager: null-check annotations before registration
- BaseDataEntity: remove duplicate id field causing Gson serialization conflict
- validateJarFile: don't require plugin.yml (modules aren't Bukkit plugins)
- SecurityPolicy: add com.ultikits.plugins to trusted packages
- plugin.yml: add adventure-platform-bukkit and HikariCP to libraries

Plugin fixes:
- Config float->double migration (YAML returns Double, not Float)
- Add (float) casts at Bukkit API boundaries (Location, playSound)
- Add plugin.yml metadata to all 9 modules for proper name/version loading
- InventoryIsolationService: add @PostConstruct to init() method

+++ plugins/UltiEssentials/src/main/java/com/ultikits/plugins/essentials/config/LobbyConfig.java
@@ -29,10 +29,10 @@ public class LobbyConfig extends AbstractConfigEntity {
     private double z = 0.0;
 
     @ConfigEntry(path = "lobby.location.yaw", comment = "水平朝向")
-    private float yaw = 0.0f;
+    private double yaw = 0.0;
 
     @ConfigEntry(path = "lobby.location.pitch", comment = "垂直朝向")
-    private float pitch = 0.0f;
+    private double pitch = 0.0;
 
     public LobbyConfig() {
         super("config/lobby.yml");
@@ -46,7 +46,7 @@ public class LobbyConfig extends AbstractConfigEntity {
     public Location getLobbyLocation() {
         return new Location(
                 Bukkit.getWorld(world),
-                x, y, z, yaw, pitch
+                x, y, z, (float) yaw, (float) pitch
         );
     }
